+++ packages/expo-file-system/src/index.ts
@@ -2,6 +2,7 @@ export * from './FileSystem';
 
 export {
   type FileCreateOptions,
+  type FilePickerOptions,
   type DirectoryCreateOptions,
   type FileHandle,
   type FileInfo,